Refactor: modal 디렉토리 생성 및 파일구조 변경에 따른 수정
